Deleted a file directory and created 404 page
Deleted the file directory for Digital Arts Showcase project and used a direct link to the website in the project modal.

Created a 404 page with a link back to index.  Updated .htaccess file.  Made sure all links work.  Updated CSS for 404 page.

Ran a fresh install of the entire website on the live server.